requirements.txt
flask
flask-cors
pymongo
google-generativeai
gunicorn